fix: export "makeExecutableSchema" through apollo-link-scalars
If there is a difference in the versions of "graphql" used in "apollo-link-scalars" and the project
including it there might be difference in the structure of the sch ema instance created by
`makeExecutableSchema`. This is the case with "graphql" 14.x and 15.x. And will result in a
TypeError - "The types returned by 'getQueryType()' are incompatible between these types.". To fix
this I am exposing `makeExecutableSchema` through "apollo-link-scalars" to ensure that the types in
the generated schema will be compatible by those expected from `withScalars`.